requirements.txt
openai
flask
python-dotenv
pyodbc==4.0.34